requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==5.1.5
django-browser-reload==1.18.0
django-decouple==2.1
django-htmx==1.21.0
django-maintenance-mode==0.21.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
futures==3.0.5
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
goslate==1.5.4
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
humanize==4.11.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.5
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-fsutil==0.14.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.3.0
uuid==1.30
whitenoise==6.8.2
yarl==1.18.3
